fix isValidAnswer() improperly checking strings
I convert all `userInput`` into a number. If this
number was anything except 0, 1, or 2, I would
return back to the `userInput` prompt.

`isValidAnswer()` was properly returning false
for all values less than 0 and greater than 2.
If a user inputted a string, it would be converted
into a number (`NaN``) as excepted.

What I failed to understand is when `userInput` is
`NaN`, (userInput == NaN) does not return true.
After some research, I discovered that
`Number.isNaN(userInput)` is the expression that
I want. This returns true when userInput is NaN.