Redesign the playRound() algorithm
In this redesign, I scrapped the nested
conditional statements checking each `userChoice`
and `computerChoice` used to determine the
winner.

Instead, I declared a new variable named
`choiceLengthDifference` which represents the
numerical difference between lengths of
`userChoice` and `computerChoice`. I then used
this variable to determine the winner of each
round.

- Is `choiceLengthDifference` === 0? Tie Game.
- Is `choiceLengthDifference` === -4? User won
with rock and computer lost with scissors
(4 - 8 = -4).
- and so on...